notification-microservice/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
email-validator==2.1.0
google-api-python-client==2.106.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0 